@@ -3,19 +3,21 @@
   "version": "0.0.0",
   "private": true,
   "dependencies": {
-    "@angular/animations": "^10.2.5",
+    "@angular/animations": "^11.2.14",
     "@angular/cdk": "^10.2.7",
-    "@angular/common": "10.2.5",
-    "@angular/compiler": "10.2.5",
-    "@angular/core": "10.2.5",
-    "@angular/forms": "10.2.5",
-    "@angular/localize": "^10.2.5",
+    "@angular/common": "11.2.14",
+    "@angular/compiler": "11.2.14",
+    "@angular/core": "11.2.14",
+    "@angular/forms": "11.2.14",
+    "@angular/localize": "^11.2.14",
     "@angular/material": "^10.2.7",
-    "@angular/platform-browser": "10.2.5",
-    "@angular/platform-browser-dynamic": "10.2.5",
-    "@angular/router": "10.2.5",
+    "@angular/platform-browser": "11.2.14",
+    "@angular/platform-browser-dynamic": "11.2.14",
+    "@angular/router": "11.2.14",
+    "@fortawesome/angular-fontawesome": "^0.8.0",
+    "@fortawesome/fontawesome-svg-core": "^6.1.1",
+    "@fortawesome/free-solid-svg-icons": "^6.1.1",
     "@wca/helpers": "^1.1.0",
-    "angular-font-awesome": "^3.1.2",
     "core-js": "2.5.7",
     "file-saver": "1.3.8",
     "font-awesome": "^4.7.0",
@@ -23,7 +25,7 @@
     "jszip": "^3.7.1",
     "ng2-cookies": "^1.0.12",
     "rxjs": "6.6.7",
-    "tslib": "^1.14.1",
+    "tslib": "^2.0.0",
     "zone.js": "0.10.3"
   },
   "scripts": {
@@ -35,25 +37,25 @@
     "e2e": "ng e2e"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "0.1002.4",
-    "@angular/cli": "~10.2.4",
-    "@angular/compiler-cli": "^10.2.5",
-    "@angular/language-service": "^10.2.5",
-    "@types/jasmine": "~2.8.6",
+    "@angular-devkit/build-angular": "0.1102.19",
+    "@angular/cli": "~11.2.19",
+    "@angular/compiler-cli": "^11.2.14",
+    "@angular/language-service": "^11.2.14",
+    "@types/jasmine": "~3.6.0",
     "@types/jasminewd2": "~2.0.3",
     "@types/jquery": "^3.3.29",
     "@types/node": "~17.0.23",
-    "codelyzer": "~6.0.2",
-    "jasmine-core": "~2.99.1",
-    "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~1.7.1",
-    "karma-chrome-launcher": "~2.2.0",
+    "codelyzer": "^6.0.0",
+    "jasmine-core": "~3.6.0",
+    "jasmine-spec-reporter": "~5.0.0",
+    "karma": "~6.3.17",
+    "karma-chrome-launcher": "~3.1.0",
     "karma-coverage-istanbul-reporter": "~2.0.0",
-    "karma-jasmine": "~1.1.1",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "~5.3.0",
+    "karma-jasmine": "~4.0.0",
+    "karma-jasmine-html-reporter": "^1.5.0",
+    "protractor": "~7.0.0",
     "ts-node": "~5.0.1",
-    "tslint": "~5.9.1",
+    "tslint": "~6.1.3",
     "typescript": "~4.0.8"
   }
 }
